refactor: modified reports management
